requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-filter>=23.0
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0
python-decouple>=3.8
